requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bitsandbytes==0.46.0
bleach==6.2.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
contourpy==1.3.2
cut-cross-entropy==25.1.1
cycler==0.12.1
dashscope==1.23.4
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
diffusers==0.33.1
dill==0.3.8
distro==1.9.0
docopt==0.6.2
docstring_parser==0.16
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
faiss-cpu==1.11.0
fastjsonschema==2.21.1
filelock==3.13.1
Flask==3.1.1
fonttools==4.58.2
frozenlist==1.7.0
fsspec==2024.6.1
gensim==4.3.3
grip==4.6.2
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.3
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
lightgbm==4.6.0
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
msgspec==0.19.0
multidict==6.4.4
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
nvitop==1.5.1
openai==1.86.0
opencv-python==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
path-and-address==2.0.1
peft==0.15.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pydantic==2.11.6
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.13.1
seaborn==0.13.2
selenium==4.33.0
sentencepiece==0.2.0
shtab==1.7.2
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.3
sympy==1.13.3
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.1+cu126
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
trio==0.30.0
trio-websocket==0.12.2
triton==3.3.0
trl==0.15.2
typeguard==4.4.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tyro==0.9.24
tzdata==2025.2
unsloth==2025.6.5
unsloth_zoo==2025.6.4
urllib3==2.4.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wordcloud==1.9.4
wrapt==1.17.2
wsproto==1.2.0
xformers==0.0.30
xxhash==3.5.0
yarg==0.1.9
yarl==1.20.1
zipp==3.23.0
